requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
confuse==1.0.0
Flask==1.1.1
Flask-MySQL==1.5.1
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mysql==0.0.2
mysqlclient==1.4.6
PyMySQL==0.9.3
PyYAML==5.3
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.0
